Add code for BuilderGraphWithRtree.py , now it'can create weighted graph.
